requirements/base.txt
Pillow==10.1.0  # https://github.com/python-pillow/Pillow

# Django
# ------------------------------------------------------------------------------
django==5.0.3  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
# DRF-spectacular for api documentation
drf-spectacular==0.26.5  # https://github.com/tfranzel/drf-spectacular
# Django CORS headers
django-cors-headers==4.3.1  # https://github.com/adamchainz/django-cors-headers
# Added By Fares
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
idna==3.6
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
stripe==10.12.0
django_filter==24.3
channels==4.1.0
channels_redis==4.2.0
gunicorn==23.0.0
daphne==4.1.2
channels[daphne]
celery==5.4.0  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.7.0  # https://github.com/celery/django-celery-beat
django-debug-toolbar==4.4.6
pyfcm==2.0.7  # https://github.com/olucurious/PyFCM
